requirements.txt
paramiko>=3.4.0
python-telegram-bot>=21.0
pyyaml>=6.0
